fix(ecommerce): corregir BIGINT en usp_store_product_getimages
MediaAssetId es BIGINT pero RETURNS TABLE declaraba INT.
Mismo patron que WaitTicketId/SaleTicketId — corregido en
13_otros.sql y usp_ecommerce.sql.

+++ web/api/sqlweb-pg/13_otros.sql
@@ -1994,7 +1994,7 @@ $function$
 -- usp_store_product_getimages
 DROP FUNCTION IF EXISTS public.usp_store_product_getimages(integer, integer, character varying) CASCADE;
 CREATE OR REPLACE FUNCTION public.usp_store_product_getimages(p_company_id integer DEFAULT 1, p_branch_id integer DEFAULT 1, p_code character varying DEFAULT NULL::character varying)
- RETURNS TABLE(id integer, url character varying, role character varying, "isPrimary" boolean, "sortOrder" integer, "altText" character varying)
+ RETURNS TABLE(id bigint, url character varying, role character varying, "isPrimary" boolean, "sortOrder" integer, "altText" character varying)
  LANGUAGE plpgsql
 AS $function$
 BEGIN
